third_party/WebKit/LayoutTests/http/tests/devtools/console/viewport-testing/console-key-links-expected.txt

Tests that console links are keyboard navigable.


Running: testNavigatingLinks
Evaluating: console.log("Text around www.chromium.org/1a multiple links, www.chromium.org/1b");console.log("www.chromium.org/2");
Message count: 2
Setting focus in prompt:

Shift+Tab:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.chromium.org/2

ArrowUp:
Viewport virtual selection: 1
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:18 www.chromium.org/2

ArrowUp:
Viewport virtual selection: 0
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:18 Text around www.chromium.org/1a multiple links, www.chromium.org/1b

ArrowDown:
Viewport virtual selection: 0
activeElement: SPAN.devtools-link
active text: www.chromium.org/1a

ArrowDown:
Viewport virtual selection: 0
activeElement: SPAN.devtools-link
active text: www.chromium.org/1b

ArrowDown:
Viewport virtual selection: 1
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:18 www.chromium.org/2

ArrowDown:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.chromium.org/2
